+++ src/types.ts
@@ -70,7 +70,7 @@ export type DefinitionInt = {
     int: number | DatumParameterKey,
 }
 
-export type DefinitionField = DefinitionConstr | DefinitionBytes | DefinitionInt
+export type DefinitionField = DefinitionConstr | DefinitionBytes | DefinitionInt | DefinitionField[] | Function
 
 export type DefinitionConstr = {
     constructor: number | DatumParameterKey,